Run clippy and enforce in CI

+++ tests/stats_extlog.rs
@@ -513,7 +513,7 @@ fn test_extloggable_bucket_counter_freq_high_value() {
     xlog!(
         logger,
         FifthExternalLog {
-            floating: 10 as f32
+            floating: 10_f32
         }
     );
 
@@ -609,7 +609,7 @@ fn test_extloggable_bucket_counter_cumul_freq() {
 #[test]
 fn test_extloggable_bucket_counter_cumul_freq_high_value() {
     let (logger, mut data) = create_logger_buffer(SLOG_TEST_STATS);
-    xlog!(logger, FifthExternalLog { floating: 8 as f32 });
+    xlog!(logger, FifthExternalLog { floating: 8_f32 });
 
     // Wait for the stats logs.
     thread::sleep(time::Duration::from_secs(TEST_LOG_INTERVAL + 1));
@@ -711,7 +711,7 @@ fn test_extloggable_bucket_counter_grouped_freq() {
         SixthExternalLog {
             name: "first".to_string(),
             error: 1,
-            floating: -7 as f32
+            floating: -7_f32
         }
     );
     xlog!(
@@ -719,7 +719,7 @@ fn test_extloggable_bucket_counter_grouped_freq() {
         SixthExternalLog {
             name: "second".to_string(),
             error: 2,
-            floating: 3.7634 as f32
+            floating: 3.7634_f32
         }
     );
 
@@ -779,7 +779,7 @@ fn test_extloggable_bucket_counter_grouped_cumul_freq() {
         SixthExternalLog {
             name: "first".to_string(),
             error: 1,
-            floating: -7 as f32
+            floating: -7_f32
         }
     );
     xlog!(
@@ -787,7 +787,7 @@ fn test_extloggable_bucket_counter_grouped_cumul_freq() {
         SixthExternalLog {
             name: "second".to_string(),
             error: 2,
-            floating: 3.7634 as f32
+            floating: 3.7634_f32
         }
     );
